Convert CCR module tests to new testing framework
