--- x-pack/plugin/ccr/qa/build.gradle
@@ -1,20 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-apply plugin: 'elasticsearch.java'
-
-dependencies {
-  api project(':test:framework')
-}
-
-subprojects {
-  tasks.withType(Test).configureEach {
-    // These fail in CI but only when run as part of checkPart2 and not individually.
-    // Tracked in : https://github.com/elastic/elasticsearch/issues/66661
-    buildParams.withFipsEnabledOnly(it)
-  }
-}
