--- x-pack/plugin/ccr/qa/rest/build.gradle
@@ -1,25 +1,0 @@
-apply plugin: 'elasticsearch.legacy-yaml-rest-test'
-apply plugin: 'elasticsearch.legacy-yaml-rest-compat-test'
-
-restResources {
-  restApi {
-    include '_common', 'cluster', 'nodes', 'indices', 'index', 'info', 'ccr'
-  }
-}
-
-dependencies {
-  yamlRestTestImplementation(testArtifact(project(xpackModule('core'))))
-}
-
-testClusters.configureEach {
-  testDistribution = 'DEFAULT'
-  // Disable assertions in FollowingEngineAssertions, otherwise an AssertionError is thrown before
-  // indexing a document directly in a follower index. In a rest test we like to test the exception
-  // that is thrown in production when indexing a document directly in a follower index.
-  jvmArgs '-da:org.elasticsearch.xpack.ccr.index.engine.FollowingEngineAssertions'
-  setting 'xpack.ml.enabled', 'false'
-  setting 'xpack.security.enabled', 'true'
-  setting 'xpack.license.self_generated.type', 'trial'
-  // TODO: reduce the need for superuser here
-  user username: 'ccr-user', password: 'ccr-user-password', role: 'superuser'
-}
