+++ x-pack/plugin/ccr/src/javaRestTest/resources/follower-roles.yml
