+++ x-pack/plugin/ccr/src/javaRestTest/resources/leader-roles.yml
