+++ x-pack/plugin/ccr/src/yamlRestTest/resources/rest-api-spec/test/ccr/auto_follow.yml
