+++ x-pack/plugin/ccr/src/yamlRestTest/resources/rest-api-spec/test/ccr/follow_and_unfollow.yml
