+++ x-pack/plugin/ccr/src/yamlRestTest/resources/rest-api-spec/test/ccr/forget_follower.yml
