+++ x-pack/plugin/ccr/src/yamlRestTest/resources/rest-api-spec/test/ccr/index_directly_into_follower_index.yml
