+++ x-pack/plugin/ccr/src/yamlRestTest/resources/rest-api-spec/test/ccr/stats.yml
